Reorganize dispatch test according to library
