+++ tests/link/mlx/test_linalg/test_solve.py
@@ -6,29 +6,7 @@ import pytest
 
 import pytensor.tensor as pt
 from pytensor import config
-from pytensor.tensor._linalg.decomposition import lu
-from pytensor.tensor._linalg.decomposition.cholesky import cholesky
 from tests.link.mlx.test_basic import compare_mlx_and_py, mlx_mode
-
-
-@pytest.mark.parametrize("lower", [True, False])
-def test_mlx_cholesky(lower):
-    rng = np.random.default_rng(15)
-    n = 3
-
-    A = pt.tensor("A", shape=(n, n))
-    A_val = rng.normal(size=(n, n))
-    A_val = (A_val @ A_val.T).astype(config.floatX)
-
-    out = cholesky(A, lower=lower)
-
-    compare_mlx_and_py(
-        [A],
-        [out],
-        [A_val],
-        mlx_mode=mlx_mode,
-        assert_fn=partial(np.testing.assert_allclose, atol=1e-6, strict=True),
-    )
 
 
 @pytest.mark.parametrize("assume_a", ["gen", "pos"])
@@ -91,21 +69,4 @@ def test_mlx_SolveTriangular(lower, trans):
         assert_fn=partial(
             np.testing.assert_allclose, atol=1e-6, rtol=1e-6, strict=True
         ),
-    )
-
-
-def test_mlx_LU():
-    rng = np.random.default_rng(15)
-
-    A = pt.tensor("A", shape=(5, 5))
-    out = lu.lu(A, permute_l=False, p_indices=True)
-
-    A_val = rng.normal(size=(5, 5)).astype(config.floatX)
-
-    compare_mlx_and_py(
-        [A],
-        out,
-        [A_val],
-        mlx_mode=mlx_mode,
-        assert_fn=partial(np.testing.assert_allclose, atol=1e-6, strict=True),
     )
